Disable Helmfile upgrade check
Starting with version 0.147.0, Helmfile tries to make an upgrade check.
As we are using Helmfile in regulated environments, we don't want that.
So to disable this, we need to set the environment variable
`HELMFILE_UPGRADE_NOTICE_DISABLED` to a non-empty value.

Also add a workflow step to print the `helmfile version` output