Allow running Helm with any other user than root
Helm stores plugins and other data at a specific location. In the
Helmfile image, this location is the `/root` directory, which already
has been made accessible for other users. But in order for other users
to be able to find the installed plugins, the Helm home environment
variables need to be set to the fixed `/root` location.

This is a temporary workaround and should be removed again as soon as
https://github.com/helmfile/helmfile/pull/546 has been merged and
released.